+++ quest/technologies/ruby/LIBRARY.md
@@ -27,8 +27,12 @@
 
 - Bundler が何のためにあるのか、プログラミング初心者にわかるように説明してください
   ```
-
+  gemをgemfileの中でまとめて管理することができる。そうすることで、gemを一つずつインストールしなくても、bundleコマンドでgemfileに記載されたgemを一括でインストールできるようになる。
   ```
 - Bundler を用いて、任意の gem を1つインストールしてください
   ```gemfile
+    gem rails
+  ```
+  ``` shell
+  $ bundle
   ```
