--- quest/technologies/ruby/vending_machine/DrinkItem.rb
@@ -1,5 +1,0 @@
-class DrinkItem < Item
-  def initialize(name, price)
-    super(name, price)
-  end
-end
