+++ quest/technologies/ruby/vending_machine/Item.rb
@@ -1,5 +1,9 @@
+# frozen_string_literal: true
+
+# vendingmachine/item.rb
 class Item
   attr_accessor :name, :price
+
   def initialize(name, price)
     @name = name
     @price = price
