--- quest/technologies/ruby/vending_machine/SnackItem.rb
@@ -1,5 +1,0 @@
-class SnackItem < Item
-  def initialize(name, price)
-    super(name, price)
-  end
-end
