+++ quest/technologies/ruby/vending_machine/cup_item.rb
@@ -1,3 +1,6 @@
+# frozen_string_literal: true
+
+# vendingmachine/cup_item.rb
 class CupItem < Item
   def initialize(name, price, temperature)
     super(name, price)
