+++ quest/technologies/ruby/vending_machine/drink_item.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+# vendingmachine/drink_item.rb
+class DrinkItem < Item
+end
