+++ quest/technologies/ruby/vending_machine/snack_item.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+# vendingmachine/snack_item.rb
+class SnackItem < Item
+end
